pyproject.toml
[tool.poetry]
name = "yelpcamp-ver2"
version = "0.1.0"
description = ""
authors = ["yuki manabe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.11"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
requests = "^2.31.0"
django-debug-toolbar = "^4.3.0"
django-allauth = "^0.61.1"
djlint = "^1.34.1"
pillow = "^10.3.0"
django-cleanup = "^8.1.0"
djangorestframework = "^3.15.1"
dj-rest-auth = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
